# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Membership Management API Documentation",
        "description": "API documentation using Swagger",
        "version": "1.0"
    },
    "paths": {
...
        "/api/business": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Lấy danh sách doanh nghiệp",
                "operationId": "26dbe1770e81c119c7275e66f251f994",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên doanh nghiệp",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số lượng doanh nghiệp trên mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách doanh nghiệp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Business"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Thêm doanh nghiệp",
                "operationId": "4714f3ec814bf207176bb5264183e7a0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBusinessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "doanh nghiệp đã được tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/business/{id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Lấy thông tin doanh nghiệp",
                "operationId": "97084bfb2c9dfb011102fd498abbfc60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của doanh nghiệp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin doanh nghiệp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "doanh nghiệp không tồn tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Cập nhật doanh nghiệp",
                "operationId": "f809cda66486896031320636f03e72da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của doanh nghiệp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBusinessRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật doanh nghiệp thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Business"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "doanh nghiệp không tồn tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Xóa doanh nghiệp",
                "operationId": "6ec006a54e5457bb7fa629437a310a3a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của doanh nghiệp",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa doanh nghiệp thành công"
                    },
                    "404": {
                        "description": "doanh nghiệp không tồn tại hoặc không thể xóa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreBusinessRequest": {
                "title": "Store Business Request",
                "description": "Yêu cầu tạo mới ngành nghề",
                "required": [
                    "business_code",
                    "business_name"
                ],
                "properties": {
                    "business_code": {
                        "type": "string",
                        "maxLength": 50,
                        "example": "IT"
                    },
                    "business_name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Công nghệ thông tin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ngành liên quan đến phần mềm, phần cứng và công nghệ",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Business": {
                "title": "Business",
                "description": "Thông tin ngành nghề",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "business_code": {
                        "type": "string",
                        "example": "IT"
                    },
                    "business_name": {
                        "type": "string",
                        "example": "Công nghệ thông tin"
                    },
                    "description": {
                        "type": "string",
                        "example": "Ngành liên quan đến phần mềm, phần cứng và công nghệ"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}